# backend/console/static/swagger/swagger.yaml
openapi: 3.0.3

info:
  title: wano-island/wano
  version: 1.0.0

paths:
  /api/v1/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login_RequestBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login_200"

  /api/v1/profile:
    get:
      security:
        - accessToken: []
      responses:
        "200":
          description: Get user profile successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile_200"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /api/v1/profile/change-password:
    post:
      security:
        - accessToken: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                confirmNewPassword:
                  type: string
      responses:
        "200":
          description: Change password successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /api/v1/shows:
    get:
      security:
        - accessToken: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Retrieved shows successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetShows_200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
    post:
      security:
        - accessToken: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                kind:
                  type: string
                originalLanguage:
                  type: string
                originalTitle:
                  type: string
                originalOverview:
                  type: string
                  nullable: true
                keywords:
                  type: array
                  nullable: true
                  items:
                    type: string
                isReleased:
                  type: boolean
      responses:
        "201":
          description: Create the show successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateShow_201"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /api/v1/providers:
    get:
      tags:
        - oauth2
      security:
        - accessToken: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Get paginated providers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetOAuth2Providers_200"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

    post:
      tags:
        - oauth2
      security:
        - accessToken: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                clientID:
                  type: string
                clientSecret:
                  type: string
                redirectURL:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                isEnabled:
                  type: boolean
      responses:
        "201":
          description: Create the oauth2 provider successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOAuth2Provider_201"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

components:
  securitySchemes:
    accessToken:
      type: apiKey
      name: X-Auth-Access-Token
      in: header

  schemas:
    Response:
      type: object
      properties:
        message:
          type: string
        messageId:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid

    PaginatedResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            pagination:
              type: object
              properties:
                page:
                  type: number
                pageSize:
                  type: number
                totalRows:
                  type: number
                totalPages:
                  type: number

    Login_RequestBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    Login_200:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                accessToken:
                  type: string
                refreshToken:
                  type: string

    Profile_200:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                locale:
                  type: string
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time

    ShowDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        originalLanguage:
          type: string
        originalTitle:
          type: string
        originalOverview:
          type: string
          nullable: true
        keywords:
          type: array
          items:
            type: string
        isReleased:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateShow_201:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ShowDTO"

    GetShows_200:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ShowDTO"

    GetOAuth2Providers_200:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                isEnabled:
                  type: boolean
                createdAt:
                  type: string
                  format: date-time
                createdBy:
                  type: string

    CreateOAuth2Provider_201:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                isEnabled:
                  type: boolean
                createdAt:
                  type: string
                  format: date-time
                createdBy:
                  type: string